Refactor and move some components
