+++ frontend/src/components/chat-interface.tsx
@@ -13,8 +13,8 @@ import AgentState from "#/types/agent-state";
 import { generateAgentStateChangeEvent } from "#/services/agent-state-service";
 import { FeedbackModal } from "./feedback-modal";
 import { useScrollToBottom } from "#/hooks/use-scroll-to-bottom";
-import TypingIndicator from "./chat/typing-indicator";
-import ConfirmationButtons from "./chat/confirmation-buttons";
+import { TypingIndicator } from "./typing-indicator";
+import { ConfirmationButtons } from "./confirmation-buttons";
 import { ErrorMessage } from "./error-message";
 import { ContinueButton } from "./continue-button";
 import { ScrollToBottomButton } from "./scroll-to-bottom-button";
