+++ frontend/src/components/feedback-form.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import hotToast from "react-hot-toast";
-import ModalButton from "./buttons/modal-button";
+import ModalButton from "./modal-button";
 import { Feedback } from "#/api/open-hands.types";
 import { useSubmitFeedback } from "#/hooks/mutation/use-submit-feedback";
