+++ frontend/src/components/terminal.tsx
